\name{dummyFunction}

\alias{dummyFunction}

\title{Dummy function to pass the CRAN checks}

\description{
This is a temporary function to pass the CRAN checks.
}

\usage{
dummyFunction(x)
}

\arguments{
  \item{x}{Any type of input.}
}

\details{
As the package QCAGUI will be archived, after being merged back to the main package QCA,
this is a temporary function to pass the CRAN checks.
This was important because many of the functions from the previous version are now in the
new version 2.5 of the package QCA, and possible namespace collisions could have happened.
}

\value{
Same object x.
}

\author{
Adrian Dusa
}

\examples{
dummyFunction(1)
}


\keyword{functions}
