\name{Ebbinghaus}

\docType{data}

\alias{Ebbinghaus}

\title{Union Growth and Decline in Western Europe 1950-95}

\description{
The \bold{Ebbinghaus} data frame has 13 rows and 10 columns.

During the early post-war period, Western trade union movements grew in membership and achieved
an institutionalized role in industrial relations and politics. However, during the last decades,
many trade unions have seen their membership decline as they came increasingly under pressures
due to the social, economic and political changes. This article reviews the main structural,
cyclical and institutional factors explaining union growth and decline. Concentrating on Western
Europe, the empirical analysis compares cross-national union density data for 13 countries over
the first period (1950-75) and for 16 countries over the second, ''crisis'' period (1975-95)
}

\usage{data(Ebbinghaus)}

\format{
The dataset contains the following columns:
\tabular{ll}{
OUT \tab union decline (outcome variable) \cr
GHENT \tab country with ''Ghent-system'' (union led unemployment insurance)\cr
WORKACCESS \tab workplace access for unions \cr
NEOCORP \tab neo-corporatist institutionalization of unions \cr
CLOSHOP \tab ''closed-shop'' tradition \cr
SOCDEM \tab government participation of Social-Democratic / Socialist party \cr
INDUS \tab share of industry in dependent unemployment \cr
PUBLIC \tab share of public sector development employment \cr
UNEMPL \tab average unemployment rate \cr
INFLA \tab average inflation rate
}
}

\source{\url{http://www.compasss.org}}

\references{
Ebbinghaus, Bernhard and Visser, Jelle 2000 \emph{Trade Unions in Western Europe since
1945}. In Flora, P. and Rothenbacher, Kraus F. (eds.) The Societies of Europe Series,
London: Macmillan Reference / Palgrave, March 2000 / New York: Grove's Dictionaries / Palgrave,
July 2000 [xxii, 808 pp. and CD-ROM]
}

\keyword{datasets}
