\name{QCAGUI-internal}

\alias{.onAttach}
\alias{deMorganLoop}
\alias{eqmccLoop}
\alias{prettyTable}
\alias{print.aE}
\alias{print.deMorgan}
\alias{print.fctr}
\alias{print.pic}
\alias{print.pof}
\alias{print.qca}
\alias{print.mqca}
\alias{print.sS}
\alias{print.tt}
\alias{uninstall}
\alias{verify.data}
\alias{verify.dir.exp}
\alias{verify.eqmcc}
\alias{verify.inf.test}
\alias{verify.qca}
\alias{verify.mqca}
\alias{verify.tt}




\alias{d.AS}
\alias{d.Bas}
\alias{d.biodiversity}
\alias{d.BWB}
\alias{d.CS}
\alias{d.CZH}
\alias{d.education}
\alias{d.Emm}
\alias{d.graduate}
\alias{d.health}
\alias{d.HK}
\alias{d.HMN}
\alias{d.homeless}
\alias{d.jobsecurity}
\alias{d.Kil}
\alias{d.Kro}
\alias{d.napoleon}
\alias{d.partybans}
\alias{d.represent}
\alias{d.RS}
\alias{d.SA}
\alias{d.socialsecurity}
\alias{d.SS}
\alias{d.stakeholder}
\alias{d.transport}
\alias{d.urban}
\alias{Emme}
\alias{HarKem}
\alias{Krook}
\alias{RagStr}
\alias{Rokkan}



\title{Internal Functions}

\description{
  The above functions are internal in the QCAGUI package which are not designed to 
  be called directly by the user. All of them are used by the \code{eqmcc} function, 
  except \code{sortMatrix} which is used by \code{allExpressions}. The verification 
  and error messages have been moved to the internal functions \code{verify.data} 
  and \code{verify.tt}.
}

\keyword{internal}