\name{QCAGUI-package}
\alias{QCAGUI-package}
\alias{QCAGUI}
\docType{package}
\title{
QCAGUI
}
\description{
A platform-independent GUI (Graphical User Interface) for the QCA package in R, based on Tcl/Tk.
}
\details{
\tabular{ll}{
Package: \tab QCAGUI\cr
Version: \tab 1.1-7\cr
Date: \tab 2006/5/31\cr
Depends: \tab R (>= 2.1.0), tcltk\cr
Suggests: \tab abind, car (>= 1.0-15), effects (>= 1.0-7), foreign, grid,
lattice, lmtest, MASS, mgcv, multcomp, nlme, nnet, QCA, relimp\cr
LazyLoad: \tab no\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.r-project.org\cr
}
}
\author{This package is 99\% based on the Rcmdr package written by
John Fox <jfox@mcmaster.ca>.
Only two additional menus were adapted for the QCA package by
Adrian Dusa <adi@roda.ro>

Maintainer: Adrian Dusa <adi@roda.ro>
}
\keyword{ package }
