\name{QCAGUI-package}

\alias{QCAGUI-package}

\alias{QCAGUI}

\docType{package}

\title{
Qualitative Comparative Analysis GUI
}

\description{
Starting with version 2.5, the package QCAGUI was merged back in the main package \pkg{QCA}.
The entire namespace and code from \pkg{QCAGUI} has been moved to package \pkg{QCA}.

For details, see:

?QCA
}

\details{
    \tabular{ll}{
    Package: \tab QCAGUI\cr
    Type:    \tab Package\cr
    Version: \tab 2.5\cr
    Date:    \tab 2016-11-13\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
Department of Sociology\cr
University of Bucharest\cr
\email{dusa.adrian@unibuc.ro}

\bold{Maintainer}:\cr
Adrian Dusa
}


\keyword{package}

