\name{Rokkan}

\docType{data}

\alias{Rokkan}

\title{Divided Working-Class Movements}

\description{
The \bold{Rokkan} data frame has 16 rows and 5 columns.

Abridged from Ragin (1987):\cr
The data was used by Rokkan (1970) in his work on nation building in Western Europe.
Rokkan used a ''configurational'' approach that bears many similarities to the Boolean
approach presented in this work. His main substantive interest was the growth of mass
democracy and the emergence of different cleavage structures in Western European
polities. One outcome that interested him was the division of some working-class
movements in these countries following the Russian Revolution into internationally
oriented wings and some into nationally oriented wings. He considered the distribution
of this outcome important because of its implication for the future of working-class
mobilization (and cleavage structures in general) in Western Europe
}

\usage{data(Rokkan)}

\format{
The dataset contains the following columns:
\tabular{ccl}{
C \tab - \tab National church (vs. state allied to Roman Catholic church) \cr
R \tab - \tab Significant Roman Catholic population and Roman Catholic participation in mass education \cr
L \tab - \tab State protection of landed interests \cr
E \tab - \tab Early state \cr
S \tab - \tab Major split in working-class movement provoked by Russian Revolution (outcome variable) \cr
  \tab   \tab (yes/no)
}
}

\source{Ragin, Charles C. 1987 \emph{The Comparative Method. Moving beyond qualitative
and quantitative strategies}, Berkeley: University of California Press, pp.129}

\references{
Rokkan, Stein 1970 \emph{Citizens, Elections, Parties}, New York: McKay
}

\keyword{datasets}
