\name{Stokke}

\docType{data}

\alias{Stokke}

\title{Causal mechanisms and regime effectiveness}

\description{
The \bold{Stokke} data frame has 9 rows and 6 columns.

The purpose of introducing mechanisms in regime research, is partly to allow detailed 
examination of the various ways in which regimes may affect behaviour: this could be
coined a magnifying purpose. The second objective is the methodological one of
facilitating systematic comparison of cases by constituting them in ways that make them
sufficiently homogeneous to permit employment of available comparative techniques.
The cases presented in this data stress on ''shaming'' as a causal mechanism for 
international resource management. Shaming highlights attempts to bring about a change
in problem-related behaviour not by material rewards or punishment but by exposing 
certain practices to third parties whose opinion matters to the target of shaming
}

\usage{data(Stokke)}

\format{
The dataset contains the following columns:
\tabular{lcl}{
A \tab - \tab Advice \cr
C \tab - \tab Commitment \cr
S \tab - \tab Shadow of the future \cr
I \tab - \tab Inconvenience \cr
R \tab - \tab Reverberation \cr
Y \tab - \tab Success (outcome variable)
}
}

\source{\url{http://www.compasss.org}}

\references{
Stokke, Olav Schram 2004 \emph{Boolean Analysis, Mechanisms, and the Study of Regime Effectiveness}
in Underdal, Arild and Young, Oran R. (eds.) Methodological Challenges and Research Strategies
Dordrecht: Kluwer Academic, pp. 87-119
}

\keyword{datasets}
