\name{aboutQCAGUI}
\alias{aboutQCAGUI}

\title{About the QCAGUI Package}
\description{
  QCAGUI Version 1.2-4
}

\author{The package is 99\% based on the Rcmdr package written by
John Fox <jfox@mcmaster.ca>.
Only some additional menus were adapted for the QCA package by
Adrian Dusa <adi@roda.ro>
}

\keyword{misc}
