\name{QCAGUI-package}
\alias{QCAGUI-package}
\alias{QCAGUI}
\docType{package}
\title{
About the QCAGUI Package
}
\description{
QCAGUI is a graphical user interface (GUI) for the QCA package, derived from R Commander. Because QCA has little to do with statistics, the menus from Rcmdr were stripped down to the very basics. In crisp sets QCA, data is binary therefore it is fairly decent to treat it as categorical (1 - presence; 0 - absence). In order to ease the primary analysis (e.g. tables of frequencies) and the creation of basic graphs, this package activates some menus that are not available in Rcmdr but for factors. Users should be aware, however, that QCAGUI is not a package for statistics; Rcmdr is better for this purpose

}
\details{
\tabular{ll}{
Package: \tab QCAGUI\cr
Version: \tab 1.9-6\cr
Date: \tab 2013-03-19\cr
Depends: \tab R (>= 2.1.0), tcltk, grDevices, utils\cr
Suggests: \tab XML\cr
LazyLoad: \tab no\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.r-project.org\cr
}
}



\author{
John Fox <jfox@mcmaster.ca>, Adrian Dusa <dusadrian@unibuc.ro>
Maintainer: Adrian Dusa <dusadrian@unibuc.ro>
}
\keyword{ package }
