\name{QCAGUI-package}

\alias{QCAGUI-package}

\alias{QCAGUI}

\docType{package}

\title{
About the QCAGUI Package
}

\description{
QCAGUI is a graphical user interface (GUI) for the QCA package, derived from R Commander. Because QCA has little to do with statistics, the menus from Rcmdr were stripped down to the very basics. In crisp sets QCA, data is binary therefore it is fairly decent to treat it as categorical (1 - presence; 0 - absence). In order to ease the primary analysis (e.g. tables of frequencies) and the creation of basic graphs, this package activates some menus that are not available in Rcmdr but for factors. Users should be aware, however, that QCAGUI is not a package for statistics; Rcmdr is better for this purpose
}

\details{
    \tabular{ll}{
    Package: \tab QCAGUI\cr
    Type: \tab Package\cr
    Version: \tab 1.3-5\cr
    Date: \tab 2008-08-10\cr
    License: \tab GPL version 2 or newer\cr
    }
}

\author{
The package is 99\% based on the Rcmdr package written by John Fox <jfox@mcmaster.ca>.
Only some additional menus were adapted for the QCA package by Adrian Dusa <adi@sas.unibuc.ro>
}

\keyword{misc}
